some tidying/renaming
